Add files via upload
Program for changing temperature and light depending on what is displayed on the screen. To use run the program and drag an mp4 file to the console. Video should play with the dots light corresponding to what is displayed on the screen. Warmer colors will result in a hotter temperature and cooler colors will result in a colder temperature.